Move and simplify method
